+++ src/components/task-edit.js
@@ -1,6 +1,6 @@
 import AbstractComponent from '../components/abstract-component.js';
 import {COLORS, DAYS, MONTH_NAMES} from '../const.js';
-import {formatTime} from '../utils.js';
+import {formatTime} from '../utils/common.js';
 
 const createRepeatingDaysMarkup = (days, repeatingDays) => {
   return days
