+++ src/main.js
@@ -9,7 +9,7 @@ import TaskEditComponent from './components/task-edit.js';
 import NoTasksComponent from './components/no-tasks.js';
 import {generateFilters} from './mock/filter.js';
 import {generateTasks} from './mock/task.js';
-import {render} from './utils.js';
+import {render} from './utils/render.js';
 
 const TASK_COUNT = 40;
 const SHOWING_TASKS_COUNT_ON_START = 8;
