+++ src/utils/render.js
@@ -3,17 +3,6 @@ const RenderPosition = {
   BEFOREEND: `beforeend`,
 };
 
-const castTimeFormat = (value) => {
-  return value < 10 ? `0${value}` : String(value);
-};
-
-const formatTime = (date) => {
-  const hours = castTimeFormat(date.getHours() % 12);
-  const minutes = castTimeFormat(date.getMinutes());
-
-  return `${hours}:${minutes}`;
-};
-
 const createElement = (template) => {
   const newElement = document.createElement(`div`);
   newElement.innerHTML = template;
@@ -32,5 +21,4 @@ const render = (container, element, place = RenderPosition.BEFOREEND) => {
   }
 };
 
-export {RenderPosition, formatTime, createElement, render};
-
+export {RenderPosition, createElement, render};
